Рефакторит компоненты
